DIP - Diplomová práca/Neural net/requirements.txt
absl-py==0.8.1
astor==0.8.0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.3
cycler==0.10.0
decorator==4.4.1
future==0.18.2
gast==0.2.2
google-auth==1.7.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
grpcio==1.25.0
h5py==2.10.0
idna==2.8
imageio==2.6.1
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras-tuner==1.0.1
kiwisolver==1.2.0
Markdown==3.2.1
matplotlib==3.2.1
numpy==1.17.4
oauthlib==3.1.0
opt-einsum==3.2.0
Pillow==6.2.1
protobuf==3.11.3
pyasn1==0.4.7
pyasn1-modules==0.2.7
pyparsing==2.4.6
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22.2.post1
scipy==1.2.0
six==1.13.0
tabulate==0.8.7
tensorboard==2.0.2
tensorflow-estimator==2.0.1
tensorflow-gpu==2.0.0
tensorflow-gpu-estimator==2.1.0
termcolor==1.1.0
terminaltables==3.1.0
tqdm==4.45.0
urllib3==1.25.7
Werkzeug==1.0.1
wrapt==1.12.1
